Split tests and SubscriptionInfo
